+++ src/main/kotlin/com/github/jensim/megamanipulator/settings/types/codehost/CodeHostSettings.kt
@@ -4,7 +4,6 @@ import com.github.jensim.megamanipulator.settings.types.AuthMethod
 import com.github.jensim.megamanipulator.settings.types.AuthMethod.USERNAME_TOKEN
 import com.github.jensim.megamanipulator.settings.types.CloneType
 import com.github.jensim.megamanipulator.settings.types.ForkSetting
-import com.github.jensim.megamanipulator.settings.types.ForkSetting.PLAIN_BRANCH
 import com.github.jensim.megamanipulator.settings.types.HostWithAuth
 import com.github.jensim.megamanipulator.settings.types.HttpsOverride
 import com.github.jensim.megamanipulator.settings.types.KeepLocalRepos
@@ -14,10 +13,8 @@ sealed class CodeHostSettings
 @SuppressWarnings("LongParameterList") constructor() : HostWithAuth {
 
     abstract val codeHostType: CodeHostSettingsType
-    abstract val baseUrl: String
     abstract val httpsOverride: HttpsOverride?
     abstract val authMethod: AuthMethod
-    abstract val username: String?
     abstract val forkSetting: ForkSetting
     abstract val cloneType: CloneType
     abstract val keepLocalRepos: KeepLocalRepos?
@@ -26,10 +23,7 @@ sealed class CodeHostSettings
     internal fun validate() {
         validateBaseUrl(baseUrl)
         if (authMethod == USERNAME_TOKEN) {
-            require(!username.isNullOrEmpty()) { "$baseUrl: username is required for auth method USERNAME_PASSWORD" }
-        }
-        if (forkSetting != PLAIN_BRANCH) {
-            require(username != null) { "username is required if forkSetting is not ${PLAIN_BRANCH.name}" }
+            require(username.isNotEmpty()) { "$baseUrl: username is required for auth method USERNAME_PASSWORD" }
         }
     }
 }